Datasets/evalplus/Java/src/test/java/com/example/HumanEval_156Test.java
package com.example;



import org.junit.Test;
import static org.junit.Assert.assertEquals;

public class HumanEval_156Test {

    @Test
    public void test_0() {
        assertEquals("xix", HumanEval_156.intToMiniRoman(19));
    }

    @Test
    public void test_1() {
        assertEquals("clii", HumanEval_156.intToMiniRoman(152));
    }

    @Test
    public void test_2() {
        assertEquals("ccli", HumanEval_156.intToMiniRoman(251));
    }

    @Test
    public void test_3() {
        assertEquals("cdxxvi", HumanEval_156.intToMiniRoman(426));
    }

    @Test
    public void test_4() {
        assertEquals("d", HumanEval_156.intToMiniRoman(500));
    }

    @Test
    public void test_5() {
        assertEquals("i", HumanEval_156.intToMiniRoman(1));
    }

    @Test
    public void test_6() {
        assertEquals("iv", HumanEval_156.intToMiniRoman(4));
    }

    @Test
    public void test_7() {
        assertEquals("xliii", HumanEval_156.intToMiniRoman(43));
    }

    @Test
    public void test_8() {
        assertEquals("xc", HumanEval_156.intToMiniRoman(90));
    }

    @Test
    public void test_9() {
        assertEquals("xciv", HumanEval_156.intToMiniRoman(94));
    }

    @Test
    public void test_10() {
        assertEquals("dxxxii", HumanEval_156.intToMiniRoman(532));
    }

    @Test
    public void test_11() {
        assertEquals("cm", HumanEval_156.intToMiniRoman(900));
    }

    @Test
    public void test_12() {
        assertEquals("cmxciv", HumanEval_156.intToMiniRoman(994));
    }

    @Test
    public void test_13() {
        assertEquals("m", HumanEval_156.intToMiniRoman(1000));
    }

    // Continue writing tests similarly...

}

// End of Test
